backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
python-multipart
pydantic-settings
python-jose[cryptography]
